Split Section and Sections
