+++ src/model/exp.rs
@@ -2,6 +2,5 @@ pub mod game_activity;
 pub mod scoremap;
 pub mod time;
 
-pub use note::sentence;
-pub use scoremap::section;
-pub use section::note;
+pub use scoremap::sections::section;
+pub use section::{note, note::sentence};
